Alex

+++ drink.cpp
@@ -2,13 +2,13 @@
 
 Drink::Drink()
 {
-    IngredientsMap = QMap<Ingredients::Ingredients, int>();
+    IngredientsMap = QMap<Ingredients::Ingredients, double>();
     Trivia = QVector<QString>();
 }
 
 Drink::Drink(QString name)
 {
     Name = name;
-    IngredientsMap = QMap<Ingredients::Ingredients, int>();
+    IngredientsMap = QMap<Ingredients::Ingredients, double>();
     Trivia = QVector<QString>();
 }
